+++ pkg/plugin/sdk/v1alpha1/BUILD.bazel
@@ -6,12 +6,12 @@ go_library(
         "aspectplugins.go",
         "setup.go",
     ],
-    importpath = "aspect.build/cli/pkg/plugins/v1alpha1",
+    importpath = "aspect.build/cli/pkg/plugin/sdk/v1alpha1",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/ioutils",
-        "//pkg/plugins/v1alpha1/config",
-        "//pkg/plugins/v1alpha1/plugin",
+        "//pkg/plugin/sdk/v1alpha1/config",
+        "//pkg/plugin/sdk/v1alpha1/plugin",
         "@com_github_hashicorp_go_hclog//:go-hclog",
         "@com_github_hashicorp_go_plugin//:go-plugin",
         "@in_gopkg_yaml_v2//:yaml_v2",
