+++ pkg/plugin/sdk/v1alpha1/config/BUILD.bazel
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "config",
     srcs = ["config.go"],
-    importpath = "aspect.build/cli/pkg/plugins/v1alpha1/config",
+    importpath = "aspect.build/cli/pkg/plugin/sdk/v1alpha1/config",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/plugins/v1alpha1/plugin",
+        "//pkg/plugin/sdk/v1alpha1/plugin",
         "@com_github_hashicorp_go_plugin//:go-plugin",
     ],
 )
