+++ pkg/plugin/sdk/v1alpha1/plugin/BUILD.bazel
@@ -6,11 +6,11 @@ go_library(
         "grpc.go",
         "interface.go",
     ],
-    importpath = "aspect.build/cli/pkg/plugins/v1alpha1/plugin",
+    importpath = "aspect.build/cli/pkg/plugin/sdk/v1alpha1/plugin",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/ioutils",
-        "//pkg/plugins/v1alpha1/proto",
+        "//pkg/plugin/sdk/v1alpha1/proto",
         "//third-party/github.com/bazelbuild/bazel/src/main/java/com/google/devtools/build/lib/buildeventstream/proto",
         "@com_github_hashicorp_go_plugin//:go-plugin",
         "@com_github_manifoldco_promptui//:promptui",
