+++ pkg/plugin/sdk/v1alpha1/plugin/interface.go
