+++ pkg/plugin/sdk/v1alpha1/proto/BUILD.bazel
@@ -12,7 +12,7 @@ proto_library(
 go_proto_library(
     name = "proto_go_proto",
     compilers = ["@io_bazel_rules_go//proto:go_grpc"],
-    importpath = "aspect.build/cli/pkg/plugins/v1alpha1/proto",
+    importpath = "aspect.build/cli/pkg/plugin/sdk/v1alpha1/proto",
     proto = ":proto_proto",
     visibility = ["//visibility:public"],
     deps = ["//third-party/github.com/bazelbuild/bazel/src/main/java/com/google/devtools/build/lib/buildeventstream/proto"],
@@ -21,6 +21,6 @@ go_proto_library(
 go_library(
     name = "proto",
     embed = [":proto_go_proto"],
-    importpath = "aspect.build/cli/pkg/plugins/v1alpha1/proto",
+    importpath = "aspect.build/cli/pkg/plugin/sdk/v1alpha1/proto",
     visibility = ["//visibility:public"],
 )
