+++ pkg/plugin/sdk/v1alpha1/proto/plugin.proto
@@ -4,7 +4,7 @@ package proto;
 
 import "third-party/github.com/bazelbuild/bazel/src/main/java/com/google/devtools/build/lib/buildeventstream/proto/build_event_stream.proto";
 
-option go_package = "aspect.build/cli/pkg/plugins/v1alpha1/proto";
+option go_package = "aspect.build/cli/pkg/plugin/sdk/v1alpha1/proto";
 
 // Plugin is the service used by the Core to communicate with a Plugin instance.
 service Plugin {
